Finished rest of the body builders

+++ backend/Domain/Email/OrderProductModel.cs
@@ -5,7 +5,7 @@
         public string Name { get; set; }
         public string Category { get; set; }
         public string Company { get; set; }
-        public decimal PriceInColones { get; set; }
+        public double PriceInColones { get; set; }
         public int Amount { get; set; }
         public string ImageURL { get; set; }
     }
